feat(availability): bind every new departure to its Product Version
The version binding landed with only recurring generation resolving a version
automatically. A departure created through the admin route recorded none, so
the guarantee — a departure can always name the definition it sells — held for
the bulk path but not the manual one.

`createSlot` now resolves the product's currently published version itself, and
generation falls back to the same resolver when no override is supplied. An
explicit `productVersionId` still wins, so a caller can materialize against a
chosen version.

The version is read through a local `productVersionsRef`, the same escape hatch
`productsRef` and `productOptionsRef` already use: Inventory owns
`product_versions` and already depends on Operations, so importing it would
close a dependency cycle. This keeps the table-privacy pair at its baseline.

Doing that required touching `service-core.ts`, which had grown past the size
gate while owning four unrelated aggregates — any edit to it failed the check.
Availability rules and start times move to `service-rules.ts`, and the shared
static-availability guard to `service-product-guard.ts`. Both are pure moves
with no behaviour change; slot and closeout lifecycle stay behind, and
`availabilityService` re-exports everything as before, so no caller changes.
service-core.ts drops from 680 to 477 lines.

Verified: typecheck 35/35, verify:architecture exit 0, table-privacy no new
reach-ins, agent-quality clean, 160 operations unit tests pass.

Part of #4027
Closes #4032
