+++ packages/operations/src/availability/service-core.ts
@@ -1,35 +1,25 @@
 import {
   type AvailabilitySlot,
   availabilityCloseouts,
-  availabilityRules,
   availabilitySlots,
-  availabilityStartTimes,
 } from "@voyant-travel/availability/schema"
 import type { EventBus } from "@voyant-travel/core"
 import { RequestValidationError } from "@voyant-travel/hono"
 import { and, asc, desc, eq, getTableColumns, gte, lt, sql } from "drizzle-orm"
 import type { PostgresJsDatabase } from "drizzle-orm/postgres-js"
 import { AVAILABILITY_SLOT_CHANGED_EVENT, type AvailabilitySlotChangedEvent } from "./events.js"
-import { productOptionsRef, productsRef } from "./products-ref.js"
+import { productOptionsRef, productsRef, resolveCurrentProductVersionId } from "./products-ref.js"
+import { assertProductAllowsStaticAvailability } from "./service-product-guard.js"
 import type {
   AvailabilityCloseoutListQuery,
-  AvailabilityRuleListQuery,
   AvailabilitySlotListQuery,
-  AvailabilityStartTimeListQuery,
   CreateAvailabilityCloseoutInput,
-  CreateAvailabilityRuleInput,
   CreateAvailabilitySlotInput,
-  CreateAvailabilityStartTimeInput,
   UpdateAvailabilityCloseoutInput,
-  UpdateAvailabilityRuleInput,
   UpdateAvailabilitySlotInput,
-  UpdateAvailabilityStartTimeInput,
 } from "./service-shared.js"
 import { paginate, toDateOrNull } from "./service-shared.js"
-import {
-  assertAvailabilityRecurrenceRule,
-  assertSlotTimingAndCapacity,
-} from "./service-validation.js"
+import { assertSlotTimingAndCapacity } from "./service-validation.js"
 import { slotEndDateLocal } from "./slot-timezone.js"
 
 export type AvailabilitySlotWithEndDateLocal = AvailabilitySlot & {
@@ -77,184 +67,6 @@ async function assertSlotOptionBelongsToProduct(
       optionId: input.optionId,
     })
   }
-}
-
-const DYNAMIC_BOOKING_MODES = new Set(["open", "stay"])
-
-async function getProductBookingMode(db: PostgresJsDatabase, productId: string) {
-  const [product] = await db
-    .select({ bookingMode: productsRef.bookingMode })
-    .from(productsRef)
-    .where(eq(productsRef.id, productId))
-    .limit(1)
-
-  return product?.bookingMode ?? null
-}
-
-async function assertProductAllowsStaticAvailability(
-  db: PostgresJsDatabase,
-  productId: string,
-  kind: "slot" | "rule",
-) {
-  const bookingMode = await getProductBookingMode(db, productId)
-  if (bookingMode && DYNAMIC_BOOKING_MODES.has(bookingMode)) {
-    throw new RequestValidationError(
-      `Dynamic ${bookingMode} products cannot author static availability ${kind}s`,
-      {
-        code: "dynamic_product_static_availability",
-        productId,
-        bookingMode,
-        kind,
-      },
-    )
-  }
-}
-
-export async function listRules(db: PostgresJsDatabase, query: AvailabilityRuleListQuery) {
-  const conditions = []
-  if (query.productId) conditions.push(eq(availabilityRules.productId, query.productId))
-  if (query.optionId) conditions.push(eq(availabilityRules.optionId, query.optionId))
-  if (query.facilityId) conditions.push(eq(availabilityRules.facilityId, query.facilityId))
-  if (query.active !== undefined) conditions.push(eq(availabilityRules.active, query.active))
-
-  const where = conditions.length ? and(...conditions) : undefined
-  return paginate(
-    db
-      .select({ ...getTableColumns(availabilityRules), productName: productsRef.name })
-      .from(availabilityRules)
-      .leftJoin(productsRef, eq(availabilityRules.productId, productsRef.id))
-      .where(where)
-      .limit(query.limit)
-      .offset(query.offset)
-      .orderBy(desc(availabilityRules.updatedAt)),
-    db.select({ count: sql<number>`count(*)::int` }).from(availabilityRules).where(where),
-    query.limit,
-    query.offset,
-  )
-}
-
-export async function getRuleById(db: PostgresJsDatabase, id: string) {
-  const [row] = await db
-    .select()
-    .from(availabilityRules)
-    .where(eq(availabilityRules.id, id))
-    .limit(1)
-  return row ?? null
-}
-
-export async function createRule(db: PostgresJsDatabase, data: CreateAvailabilityRuleInput) {
-  assertAvailabilityRecurrenceRule(data.recurrenceRule)
-
-  if (data.active !== false) {
-    await assertProductAllowsStaticAvailability(db, data.productId, "rule")
-  }
-
-  const [row] = await db.insert(availabilityRules).values(data).returning()
-  return row
-}
-
-export async function updateRule(
-  db: PostgresJsDatabase,
-  id: string,
-  data: UpdateAvailabilityRuleInput,
-) {
-  if (data.recurrenceRule !== undefined) {
-    assertAvailabilityRecurrenceRule(data.recurrenceRule)
-  }
-
-  if (data.productId !== undefined || data.active === true) {
-    const [current] = await db
-      .select({ productId: availabilityRules.productId, active: availabilityRules.active })
-      .from(availabilityRules)
-      .where(eq(availabilityRules.id, id))
-      .limit(1)
-    if (!current) return null
-
-    const nextProductId = data.productId ?? current.productId
-    const nextActive = data.active ?? current.active
-    if (nextActive) {
-      await assertProductAllowsStaticAvailability(db, nextProductId, "rule")
-    }
-  }
-
-  const [row] = await db
-    .update(availabilityRules)
-    .set({ ...data, updatedAt: new Date() })
-    .where(eq(availabilityRules.id, id))
-    .returning()
-  return row ?? null
-}
-
-export async function deleteRule(db: PostgresJsDatabase, id: string) {
-  const [row] = await db
-    .delete(availabilityRules)
-    .where(eq(availabilityRules.id, id))
-    .returning({ id: availabilityRules.id })
-  return row ?? null
-}
-
-export async function listStartTimes(
-  db: PostgresJsDatabase,
-  query: AvailabilityStartTimeListQuery,
-) {
-  const conditions = []
-  if (query.productId) conditions.push(eq(availabilityStartTimes.productId, query.productId))
-  if (query.optionId) conditions.push(eq(availabilityStartTimes.optionId, query.optionId))
-  if (query.facilityId) conditions.push(eq(availabilityStartTimes.facilityId, query.facilityId))
-  if (query.active !== undefined) conditions.push(eq(availabilityStartTimes.active, query.active))
-  const where = conditions.length ? and(...conditions) : undefined
-
-  return paginate(
-    db
-      .select({ ...getTableColumns(availabilityStartTimes), productName: productsRef.name })
-      .from(availabilityStartTimes)
-      .leftJoin(productsRef, eq(availabilityStartTimes.productId, productsRef.id))
-      .where(where)
-      .limit(query.limit)
-      .offset(query.offset)
-      .orderBy(availabilityStartTimes.sortOrder, availabilityStartTimes.createdAt),
-    db.select({ count: sql<number>`count(*)::int` }).from(availabilityStartTimes).where(where),
-    query.limit,
-    query.offset,
-  )
-}
-
-export async function getStartTimeById(db: PostgresJsDatabase, id: string) {
-  const [row] = await db
-    .select()
-    .from(availabilityStartTimes)
-    .where(eq(availabilityStartTimes.id, id))
-    .limit(1)
-  return row ?? null
-}
-
-export async function createStartTime(
-  db: PostgresJsDatabase,
-  data: CreateAvailabilityStartTimeInput,
-) {
-  const [row] = await db.insert(availabilityStartTimes).values(data).returning()
-  return row
-}
-
-export async function updateStartTime(
-  db: PostgresJsDatabase,
-  id: string,
-  data: UpdateAvailabilityStartTimeInput,
-) {
-  const [row] = await db
-    .update(availabilityStartTimes)
-    .set({ ...data, updatedAt: new Date() })
-    .where(eq(availabilityStartTimes.id, id))
-    .returning()
-  return row ?? null
-}
-
-export async function deleteStartTime(db: PostgresJsDatabase, id: string) {
-  const [row] = await db
-    .delete(availabilityStartTimes)
-    .where(eq(availabilityStartTimes.id, id))
-    .returning({ id: availabilityStartTimes.id })
-  return row ?? null
 }
 
 export async function listSlots(db: PostgresJsDatabase, query: AvailabilitySlotListQuery) {
@@ -362,10 +174,18 @@ export async function createSlot(
       ? data.initialPax
       : data.remainingPax
 
+  // Bind the departure to the Product definition it is created from, so later
+  // Product edits cannot silently rewrite what it sells (#4032). An explicit
+  // value wins — that is how a caller materializes against a chosen version
+  // rather than whatever is currently published.
+  const productVersionId =
+    data.productVersionId ?? (await resolveCurrentProductVersionId(db, data.productId))
+
   const [row] = await db
     .insert(availabilitySlots)
     .values({
       ...data,
+      productVersionId,
       remainingPax,
       startsAt: new Date(data.startsAt),
       endsAt: toDateOrNull(data.endsAt),
